// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.15"
  }
  public: {
    Tables: {
      collections: {
        Row: {
          created_at: string
          handle: string
          position: number
          title: string
        }
        Insert: {
          created_at?: string
          handle: string
          position?: number
          title: string
        }
        Update: {
          created_at?: string
          handle?: string
          position?: number
          title?: string
        }
        Relationships: []
      }
      product_collections: {
        Row: {
          collection_handle: string
          product_handle: string
        }
        Insert: {
          collection_handle: string
          product_handle: string
        }
        Update: {
          collection_handle?: string
          product_handle?: string
        }
        Relationships: [
          {
            foreignKeyName: "product_collections_collection_handle_fkey"
            columns: ["collection_handle"]
            isOneToOne: false
            referencedRelation: "collections"
            referencedColumns: ["handle"]
          },
          {
            foreignKeyName: "product_collections_product_handle_fkey"
            columns: ["product_handle"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["handle"]
          },
        ]
      }
      products: {
        Row: {
          brand: string
          compare_at_price: number | null
          created_at: string
          description: string
          handle: string
          images: string[]
          price: number
          rating: number
          reviews: number
          sizes: string[]
          title: string
        }
        Insert: {
          brand?: string
          compare_at_price?: number | null
          created_at?: string
          description?: string
          handle: string
          images?: string[]
          price?: number
          rating?: number
          reviews?: number
          sizes?: string[]
          title: string
        }
        Update: {
          brand?: string
          compare_at_price?: number | null
          created_at?: string
          description?: string
          handle?: string
          images?: string[]
          price?: number
          rating?: number
          reviews?: number
          sizes?: string[]
          title?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never